aigis-backend/pyproject.toml
[project]
name = "aigis-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-deepseek>=0.1.4",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
